+++ widgets/WikiLinkJsModal.php
@@ -2,12 +2,12 @@
 
 namespace humhub\modules\wiki\widgets;
 
-use humhub\widgets\Button;
-use humhub\widgets\Modal;
-use humhub\widgets\ModalButton;
+use humhub\widgets\bootstrap\Button;
+use humhub\widgets\modal\JsModal;
+use humhub\widgets\modal\ModalButton;
 use Yii;
 
-class WikiLinkModal extends Modal
+class WikiLinkJsModal extends JsModal
 {
     public $id = 'wikiLinkModal';
 
@@ -18,7 +18,7 @@ class WikiLinkModal extends Modal
 
     public function init()
     {
-        $this->header = Yii::t('WikiModule.base', '<strong>Wiki</strong> link');
+        $this->title = Yii::t('WikiModule.base', '<strong>Wiki</strong> link');
         $this->body = $this->render('wikiLinkModal', ['contentContainer' => $this->contentContainer]);
         $this->footer = Button::save()->action('wiki.linkExtension.setEditorLink')->loader(false) . ModalButton::cancel();
         parent::init();