+++ widgets/views/menu.php
@@ -5,7 +5,7 @@
  * @license https://www.humhub.com/licences
  */
 
-use humhub\libs\Html;
+use humhub\helpers\Html;
 use humhub\modules\ui\menu\MenuEntry;
 use humhub\modules\wiki\widgets\WikiMenu;
 
@@ -20,12 +20,11 @@ use humhub\modules\wiki\widgets\WikiMenu;
 
     <?php if (!empty($entries)) : ?>
         <?= Html::beginTag('div', $options) ?>
-            <button type="button" class="btn btn-info btn-sm dropdown-toggle" data-toggle="dropdown" aria-haspopup="true">
-                <span class="caret"></span>
+            <button type="button" class="btn btn-accent btn-sm dropdown-toggle" data-bs-toggle="dropdown" aria-haspopup="true">
             </button>
-            <ul class="dropdown-menu pull-right">
+            <ul class="dropdown-menu dropdown-menu-end">
                 <?php foreach ($entries as $entry) : ?>
-                    <li><?= $entry->render() ?></li>
+                    <li><?= $entry->render(['class' => 'dropdown-item']) ?></li>
                 <?php endforeach; ?>
             </ul>
         <?= Html::endTag('div')?>